spotify_data_collect/requirements.txt
flask
psycopg2-binary
spotipy
